조회수 증가를 ajax로 실행

+++ src/main/webapp/WEB-INF/jsp/usr/article/detail.jsp
@@ -1,10 +1,32 @@
-
 <%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
 <%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c"%>
 
 <c:set var="pageTitle" value="게시물 내용" />
 <%@ include file="../common/head.jspf"%>
 
+<script>
+  const params = {};
+  params.id = parseInt('${param.id}');
+</script>
+
+<script>
+  function ArticleDetail__increaseHitCount() {
+    $.get('../article/doIncreaseHitCountRd', {
+      id : params.id,
+      ajaxMode: 'Y'
+    }, function(data) {
+      $('.article-detail__hit-count').empty().html(data.data1);
+    }, 'json');
+  }
+  
+  $(function() {
+    // 실전코드
+    // ArticleDetail__increaseHitCount();
+    
+    // 임시코드
+    setTimeout(ArticleDetail__increaseHitCount, 500);
+  })
+</script>
 
 <section class="mt-5">
   <div class="container mx-auto px-3">
@@ -16,7 +38,9 @@
         <tbody>
           <tr>
             <th>번호</th>
-            <td><div class="badge badge-primary">${article.id}</div></td>
+            <td>
+              <div class="badge badge-primary">${article.id}</div>
+            </td>
           </tr>
           <tr>
             <th>작성날짜</th>
@@ -32,7 +56,9 @@
           </tr>
           <tr>
             <th>조회수</th>
-            <td>${article.hitCount}</td>
+            <td>
+              <span class="badge badge-primary article-detail__hit-count">${article.hitCount}</span>
+            </td>
           </tr>
           <tr>
             <th>제목</th>
@@ -52,10 +78,14 @@
         <a class="btn btn-link" href="../article/modify?id=${article.id}">게시물 수정</a>
       </c:if>
       <c:if test="${article.extra__actorCanDelete}">
-        <a class="btn btn-link" onclick="if ( confirm('정말 삭제하시겠습니까?') == false ) return false;"
-          href="../article/doDelete?id=${article.id}">게시물 삭제</a>
+        <a class="btn btn-link" onclick="if ( confirm('정말 삭제하시겠습니까?') == false ) return false;" href="../article/doDelete?id=${article.id}">게시물 삭제</a>
       </c:if>
     </div>
   </div>
 </section>
+
+<!--
+<iframe src="http://localhost:8011/usr/article/doIncreaseHitCountRd?id=2" frameborder="0"></iframe>
+-->
+
 <%@ include file="../common/foot.jspf"%>
